Making unit tests happier.
