+++ csvcubed/csvcubed/readers/cubeconfig/v1_0/__init__.py
@@ -1,1 +1,1 @@
-import configdeserialiser
+from . import configdeserialiser
